HoraireFragment: Give the possibility to choose a schedule when exporting
When exporting a schedule to a calendar all the given
schedules exposed get exported to the user's calendar
application. This might be an unwanted behavior
as the user might not want all of them. (For example, he/
she only wants the courses)

Give the choice to the user to select which calendar type
he/she wants.

At the same time, refactor the single AsyncTask for
HoraireManger as an inner class. nnerclass.