EvaluationCours/Helper: Remove getDateTimeFormat()
The generation of the new ETS WSDL definition was based on the
fact that the Java 8 Time API would be used. However, the ones
used in ETSMobile are based around JodaTime.

For this reason, since we already had adapted the classes
beforehand, simply return the correct DateTime object when
obtaining a property.